requirements.txt
streamlit
groq
sentence-transformers
chromadb
pypdf
duckduckgo-search
torch
python-dotenv
